requirements.txt
opencv-contrib-python>=4.8.0
numpy>=1.24.0
Pillow>=9.0.0